Remaining DB methods need to be implemented in BusinessModel #7
Added remaining functionality and refactored existing methods
Added check for record existence in Data.lock - throws RecordNotFoundException if it doesn't exist
Organised imports

+++ src/suncertify/db/Data.java
@@ -68,6 +68,10 @@ public class Data implements DB {
     public void delete(int recNo, long lockCookie)
             throws RecordNotFoundException, SecurityException {
         
+        if (!database.recordExists(recNo)) {
+            throw new RecordNotFoundException("Record number " + recNo + " does not exist");
+        }
+        
         if (isValidCookie(recNo, lockCookie)) {
             database.deleteRecord(recNo); 
         }
